+++ server/routes/index.js
@@ -3,12 +3,6 @@ const loginUser = require('../controller/Login');
 const registerUser = require('../controller/register');
 const router = express.Router();
 
-router.get("/home", (req, res) => {
-  res.send("something nicepppp");
-});
-
-
-
 router.post('/login', loginUser);
 router.post('/signup', registerUser )
